+++ lib/generators/react_on_rails/install_generator.rb
@@ -17,6 +17,13 @@ module ReactOnRails
                    desc: "Install Redux gems and Redux version of Hello World Example. Default: false",
                    aliases: "-R"
 
+      # --redux
+      class_option :node,
+                   type: :boolean,
+                   default: false,
+                   desc: "Sets up node as a server rendering option. Default: false",
+                   aliases: "-N"
+
       # --ignore-warnings
       class_option :ignore_warnings,
                    type: :boolean,
@@ -46,6 +53,7 @@ module ReactOnRails
         invoke "react_on_rails:base"
         invoke "react_on_rails:react_no_redux" unless options.redux?
         invoke "react_on_rails:react_with_redux" if options.redux?
+        invoke "react_on_rails:node" if options.node?
       end
 
       # NOTE: other requirements for existing files such as .gitignore or application.